backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standarc,standard]>=0.115.12",
    "openai>=1.70.0",
    "pytest-playwright>=0.7.0",
    "ruff>=0.11.2",
    "python-dotenv>=1.0.0",
]
